Adds RubyConf#in_progress? method and spec
The spec in this commit is flaky. It passes before (and after)
RubyConf, but fails during the event.

Looking at the spec, you might think the author clearly missed
the specs for what happens during and after the conference. But,
this is exaclty what many time-centric specs do. They cover the
current context, and completely neglect Daylight Saving Time.
So, when the clock finally changes time zones (or when the
event begins) those tests will fail.

Co-authored-by: Fito von Zastrow <adolfovon@gmail.com>