Fixes flaky spec for RubyConf#session_description
This spec will now pass even without connectivity to the server.

A familiar response I see to this kind of test is the question,
"How will we know if the API changes?" This is a perfectly valid
question — for a different test.

This spec is a unit test for our code. It tests that we request
and parse the data correctly, and that we return the correct
description. Testing that the API hasn't changed is more of an
integration test that should be a part of a different suite.
It should also probably only need to run once a day.

Co-authored-by: Fito von Zastrow <adolfovon@gmail.com>