Add RubyConf#favorite_session accessor methods and specs
These methods work fine. But, the specs are order dependent. Run
in one order and they'll pass. Run in the opposite order, and one
will fail due to the shared state of the cache being polluted by
the other spec.

Run `bundle exec rspec --order random:1` to reproduce the failure.

Co-authored-by: Fito von Zastrow <adolfovon@gmail.com>