Fixes order dependency flakiness for RubyConf#favorite_session
This commit uses dependency injection to isolate each from the
other by giving each test their own copy of the cache.

Note that in adding the depedency injection, it is important to
maintain the existing interface of your class so that you don't
have to grep the codebase and change every reference to the class.
You can do this by adding a default value for the injected
parameter.

Co-authored-by: Fito von Zastrow <adolfovon@gmail.com>