Adds RubyConf#reserve_seat and RubyConf#attendees methods and specs
These methods store state in a text file on the local file system.
The specs work fine when run locally, in any order. But, when run in
parallel, the ansynchronous nature of File IO allows the tests to
interfere with one another.

Run `bundle exec prspec spec` to run the specs in parallel. This
is the only way to see them fail.

Co-authored-by: Fito von Zastrow <adolfovon@gmail.com>